pyproject.toml
[project]
name = "sumplete"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Markushik", email = "zemly@duck.com" }
]
dependencies = [
    "winloop>=0.1.6",
    "aiogram-dialog>=2.1.0",
    "dishka>=1.3.0",
    "adaptix>=3.0.0b7",
    "dynaconf>=3.2.6",
    "redis>=5.0.8",
    "sqlalchemy>=2.0.32",
    "pendulum>=3.0.0",
    "fluent-runtime>=0.4.0",
    "psycopg[binary]>=3.2.1",
    "numba>=0.60.0",
    "numpy==2.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6.2",
    "ruff-lsp>=0.0.55",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sumplete"]
